// assets/Script/constants/enums.ts
export enum SymbolAnimation {
    Static = "static",
    Win = "win",
    Blur = "blur",
    Dark = "dark",
    Idle = "idle",
    Landing = "landing"
}

export enum SpinPhase {
    Stop = "stop",
    StartBounce = "startBounce",
    Accelerate = "accelerate",
    Spinning = "spinning",
    Slowdown = "slowdown",
    FinishBounce = "finishBounce"
}

export enum SlotEvents {
    OnReelStop = "onReelStop",
    OnDrumFinish = "onDrumFinish"
}

export enum WinType {
    Regular = "regular",
    BigWin = "bigwin",
    EpicWin = "epicWin"
}